The search view design was updated to how it is displayed in explore
